+++ gauss.h
@@ -2,9 +2,9 @@
 #define H_GAUSS
 
 
-void gauss(double *tab, double *res, int N);
-void gauss_inff(double *tab, double *res, int N);
-void gauss_supp(double *tab, double *res, int N);
-void gauss_res(double *tab, double *res, int N);
+void gauss(float *tab, float *res, int N);
+void gauss_inff(float *tab, float *res, int N);
+void gauss_supp(float *tab, float *res, int N);
+void gauss_res(float *tab, float *res, int N);
 
 #endif
