+++ testMatrice.h
@@ -2,18 +2,18 @@
 #define H_TEST_MATRICE
 
 int min(int a, int b);
-void init_zero(double tab[]);
-void a_bord1(double tab[]);
-void a_bord2(double tab[]);
-void ding_dong(double tab[]);
-void franc(double tab[]);
-void hilbert(double tab[]);
-void hilbert2(double tab[]);
-void kms(double tab[]);
-void lehmer(double tab[]);
-void lotkin(double tab[]);
-void moler(double tab[]);
-void affiche_tab(double *tab);
-void affiche_tab_res(double *tab);
+void init_zero(float tab[]);
+void a_bord1(float tab[]);
+void a_bord2(float tab[]);
+void ding_dong(float tab[]);
+void franc(float tab[]);
+void hilbert(float tab[]);
+void hilbert2(float tab[]);
+void kms(float tab[]);
+void lehmer(float tab[]);
+void lotkin(float tab[]);
+void moler(float tab[]);
+void affiche_tab(float *tab);
+void affiche_tab_res(float *tab);
 
 #endif
